Moved EIP712 filtering related code into its own file
